requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
python-pptx>=0.6.21
pymupdf>=1.22.5
pillow>=10.0.0
markdown>=3.4.3
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0 